Put timer logic in timer.js
