+++ client/src/App.js
@@ -1,10 +1,10 @@
+import React, {useContext, useState} from 'react';
 import './App.css';
 import DayVotingMenu from './components/DayVotingMenu';
 import { GameContext } from './GameContext';
-import React, {useContext, useState} from 'react';
 
 function App() {
-  const { players, setPlayers } = useContext(GameContext)
+  const { players } = useContext(GameContext)
   const [isVotingTime, setIsVotingTime] = useState(true)
 
   return (
