+++ client/src/components/DayVotingMenu.js
@@ -11,30 +11,7 @@ const DayVotingMenu = ({ currentVoter, setIsVotingTime }) => {
             votingTime } = useContext(GameContext)
     const [time, setTime] = useState(votingTime)
 
-    let timerInterval;
 
-    const decrementTimer = () => {
-        if (time > 0) {
-            setTime(time - 1)
-        } else {
-            clearInterval(timerInterval)
-            players.forEach((player) => {
-                if (player.alive && document.getElementById("radio" + player.name).checked) {
-                    let newPlayers = [...players];
-                    //console.log("incrementing " + player.name + " from " + player.currentVotes + " to " + (player.currentVotes + 1))
-                    newPlayers[player.id].currentVotes++;
-                    newPlayers[currentVoter.id].hasVoted = true;
-                    setPlayers(newPlayers);
-                }
-            }) 
-            
-            setIsVotingTime(false)
-        }
-    }
-    
-    const startSetInterval = () => {
-        timerInterval = setInterval(decrementTimer, 1000);
-    }
 
     const confirmVote = () => {
         
@@ -47,7 +24,6 @@ const DayVotingMenu = ({ currentVoter, setIsVotingTime }) => {
         players.forEach((player) => {
             if (player.alive && document.getElementById("radio" + player.name).checked) {
                 let newPlayers = [...players];
-                //console.log("incrementing " + player.name + " from " + player.currentVotes + " to " + (player.currentVotes + 1))
                 newPlayers[player.id].currentVotes++;
                 newPlayers[currentVoter.id].hasVoted = true;
                 setPlayers(newPlayers);
@@ -59,14 +35,12 @@ const DayVotingMenu = ({ currentVoter, setIsVotingTime }) => {
             setIsVotingTime(false);
         }
     }
-
-    startSetInterval();
     
     return (
         <div className="dayVoting">
             <div className="header">
                 <h1 className="title">Day {dayNum}: Voting</h1>
-                <Timer time={time} setTimer={setTime}/>
+                <Timer time={time} setTime={setTime} setIsVotingTime={setIsVotingTime} currentVoter={currentVoter}/>
             </div>
             <h3 className="instructions">Vote for a person to kill, voting ends in {time} seconds.</h3>
             <div className="actionPanel">
