+++ client/src/components/Timer.js
@@ -1,7 +1,35 @@
+import React, {useContext} from 'react'
+import { GameContext } from '../GameContext'
 
+const Timer = ({ time, setTime, setIsVotingTime, currentVoter }) => {
+    const { players, setPlayers } = useContext(GameContext);
 
-const Timer = ({ time }) => {
+    let timerInterval;
+
+    const decrementTimer = () => {
+        if (time > 0) {
+            setTime(time - 1)
+        } else {
+            clearInterval(timerInterval)
+            players.forEach((player) => {
+                if (player.alive && document.getElementById("radio" + player.name).checked) {
+                    let newPlayers = [...players];
+                    newPlayers[player.id].currentVotes++;
+                    newPlayers[currentVoter.id].hasVoted = true;
+                    setPlayers(newPlayers);
+                }
+            }) 
+            
+            setIsVotingTime(false)
+        }
+    }
     
+    const startSetInterval = () => {
+        timerInterval = setInterval(decrementTimer, 1000);
+    }
+
+    startSetInterval();
+
     return (
         <div>
             <h1>{time}</h1>
